Add functionality to the operators' buttons
Make the calculator work. Retrieve the components of the operation from the variables where they are stored and call the `operate` method using these values.

Reset operation variables and clear the display when
the user presses the clear button to make sure the user starts fresh.

The calculator should not evaluate more than a single
pair of numbers at a time. If a user enters a second operator,
First, evaluate the initial pair of numbers, then display the result of
That calculation. Finally, use that result as the first number in
a new calculation, along with the next operator.