Fix considering the result as the first operand
We consider the result of a previous operation as the
first operand for the next one.
The problem was that we set the isFirstOperand variable to false
after that, and this makes the program expect a second operand instead
of an operator

So to solve it, we just let the isFirstOperand variable be true.